+++ .github/workflows/build_recipes.yaml
@@ -92,7 +92,7 @@ jobs:
         shell: bash -el {0}
         if: github.event_name == 'push'
         run: |
-          python emci build changed $GITHUB_WORKSPACE/ origin/main~1 origin/main
+          python -m emci build changed $GITHUB_WORKSPACE/ origin/main~1 origin/main
 
       ################################################################
       # PULL_REQUEST
@@ -101,7 +101,7 @@ jobs:
         shell: bash -el {0}
         if: github.event_name == 'pull_request'
         run: |
-          python emci build changed $GITHUB_WORKSPACE/ origin/main HEAD
+          python -m emci build changed $GITHUB_WORKSPACE/ origin/main HEAD
 
       ################################################################
       # UPLOAD
