+++ .github/workflows/build_simple.yaml
@@ -92,7 +92,7 @@ jobs:
       - name: Build package not on emscripten-forge (advanced)
         shell: bash -el {0}
         run: |
-          python emci build explicit $GITHUB_WORKSPACE/emci/testing/test_recipes/only_py_tests --emscripten-wasm32
+          python -m emci build explicit $GITHUB_WORKSPACE/test_recipes/only_py_tests --emscripten-wasm32
         
 
       ################################################################
@@ -101,7 +101,7 @@ jobs:
       - name: Build regex
         shell: bash -el {0}
         run: |
-          python emci build explicit $GITHUB_WORKSPACE/recipes/recipes_emscripten/regex --emscripten-wasm32
+          python -m emci build explicit $GITHUB_WORKSPACE/recipes/recipes_emscripten/regex --emscripten-wasm32
 
       ################################################################
       # Build a simple package with legacy boa
