+++ emci/find_recipes_with_changes.py
@@ -1,0 +1,22 @@
+from .git_utils import find_files_with_changes
+from .constants import RECIPES_SUBDIR_MAPPING
+import os
+
+def find_recipes_with_changes(old, new):
+    files_with_changes = find_files_with_changes(old=old, new=new)
+
+    recipes_with_changes = {k: set() for k in RECIPES_SUBDIR_MAPPING.keys()}
+    # print("recipes_with_changes", recipes_with_changes)
+    for subdir in RECIPES_SUBDIR_MAPPING.keys():
+        for file_with_change in files_with_changes:
+            if file_with_change.startswith(f"recipes/{subdir}/"):
+                # print(file_with_change)
+                file_with_change = file_with_change[len(f"recipes/{subdir}/") :]
+                file_with_change = os.path.normpath(file_with_change)
+                recipe = file_with_change.split(os.sep)[0]
+                recipes_with_changes[subdir].add(recipe)
+
+    for subdir in RECIPES_SUBDIR_MAPPING.keys():
+        recipes_with_changes[subdir] = sorted(list(recipes_with_changes[subdir]))
+    return recipes_with_changes
+
