+++ emci/git_utils.py
@@ -1,0 +1,17 @@
+import subprocess
+
+def find_files_with_changes(old, new):
+    cmd = ["git", "diff", "--name-only", old, new]
+    result = subprocess.run(
+        cmd,
+        shell=False,
+        check=True,
+        capture_output=True,
+    )
+    output_str = result.stdout.decode()
+    error_str = result.stderr.decode()
+    if len(error_str):
+        print(error_str)
+
+    files_with_changes = output_str.splitlines()
+    return files_with_changes
