+++ test_recipes/only_py_tests/rattler_recipe.yaml
